[MachineCombiner] Allow reassociating of operands that are associative but not communative
The opcodes of these operations may not be identical to each other even though they are
semanitcally eqivilant. Special care must be taken to ensure the corrent opcodes stay with the
correct operands. The assumption is made that the operation should be associated with the
second operand as this holds true in all hypothesized cases. However, I probably ought to formalize this.

For instance, in RISCV, SHL and MUL can be reassociated even though they are different operations and
SHL is not communative.

%shl = SHL %0, %1
%out = MUL killed %shl, %2

is equivliant to:
%mul = MUL %0, %2
%out = SHL %mul, %1

because SHL %0, %1 is equivilant to MUL %0, 2^^%1.

This is backend specific and each specific backend can define their own operations for