[SLPVectorizer] Cache information about external uses to better handle overlapping chains
For instance, if two chains both derive from the same vectorizable-loads, they pay a price
since the load has a use not part of their chain, so the vectorizer thinks that those loads
will have to be extracted from the vector. This price may be high enough that the chains
are deemed to not be profitable to vectorize. However, both chains would be profitble if we
knew the other one was going to be vectorized as well.

The change stores information about chains that weren't vectorized due to external uses (if
the external uses have acceptable properties), and allows future chains to reference that
information to determine if those external uses will be eventually vectorized too. In such
a case, it can bypass the external use penalties and get a better cost estimation.