[SLPVectorization] Run SLPVectorization until a fixpoint is reached.
This is important in the current implementation, in the case when two
chains overlap. One of the chains is not profitable due to the
extraction of elements needed for the other chain (which we haven't yet
calculated, so these are external uses by random scalars). When we calculate
the second chain, it is profitable despite the external uses by the first
chain. Thus, the second chain is vectorized.

Currently, we would now exit from the SLPVectorized, with the first
chain un-vectorized. However, the first chain is clearly vectorizable
now that we have eliminated those external uses (by vectorizing them into the
second chain). Fixpoint solves this issue.